topal: add ebuild for latest version
Had to tune the patch to work. Also, there are problems with ada compilers
in the tree - they don't really build.

I've worked around that by choosing gnat-gcc-4.5 which builds but doesn't
install due to bug #561968. However that one is a minor one to workaround
(add -lto to IUSE), so I decided to use that one as a dependency.