alpine-2.20: remove support for topal integration through patches
The patches have not been updated for >alpine-2.00, sending mime
multipart/signed messages with them doesn't work (or maybe it was just
missing the procmail component to correctly format the email), and
attempting to open attachments crashes alpine.

There is an alternative route for topal integration that doesn't require
patching anything and only uses topal at runtime. This is the recommended
configuration.

Fixed patches so they that apply (but don't fully work) and they will
remain in files/.