plan9port: make prefix safe
The mixture of do* and shell commands causes a bit of a headache. I
introduced a EPLAN9 variable that is prefixed, which is used when absolute
paths are necessary (cp, env file). When using do* functions the normal
PLAN9 variable is used. The end result works in prefix.

I'm not sure that replacing all occurrences of /usr/local/plan9 (as is
done in the tree ebuild) using sed is the best choice. This makes some of
the documentation host specific and doesn't add any value. It's worth
considering to go back to the moveplan9.sh script which only changes the
needed things, leaving documentation stock.