plan9port: fix bugs in ebuild
I added use flags for aqua and truetype with correct dependencies.
Freebsd-10 was being incorrectly handled in 9l, due to matching freebsd1*,
but this is fixed by a new patch. src_configure was simplified by using an
array, and explicitlyi specifies all settings. In an attempt to not use
the host compiler on prefix, the CC9 variable is exported with the toolchain
compiler. Fixes #567260.